Update patch set 1
Patch Set 1:

This needs some cleanup before it is ready to submit, but I wanted to get some feedback first.

When using t9 search in other languages we want to show both the english and alternate t9 characters.  This necessitates putting the t9 labels to the right of the digits due to vertical space constraints.  

Simplest solution is to always put the labels to the right even when we are in english-only mode (similar to cm11).  That's what I have done here.  Screenshots:

http://imgur.com/a/ZeLGO

What do people think?

Patch-set: 1