Update patch set 1
Patch Set 1:

Having the t9 labels to the right side makes the most sense. While there may be a small visual weight issue with the numbers not perfectly aligning over the call btn, that is trumped by optimizing the layout for a secondary label. Also, placing the labels on the right-hand side collapses the dialpad considerably, making it easier to dial with one hand. If the labels were under the number, there would be reachability issues on larger screens. Just my 2¢

Patch-set: 1