Update patch set 1
Patch Set 1:

> (1 comment)

isVolteAvailable checks to see if the device has volte support (via platform).  if that is set, it will show "Enhanced 4G LTE Mode" user settings which is queried by isEnhanced4gLteModeSettingEnabledByUser().  However, that is not enough to to show the Volte indicator.  isVolteCallCapable() was added to check for the volte registration that occurs from network.  so if the user loses LTE signal, it should not show volte indicator on dialer UI.

Patch-set: 1