Update patch set 1
Patch Set 1:

"  so if the user loses LTE signal, it should not show volte indicator on dialer UI. "

Makes sense. But I don't think thats how it is implemented in Dialer / InCallUI currently. isVolteInUse is called once on set-up and the label is then persistent till the UI is torn down.

Patch-set: 1