Update patch set 4
Patch Set 4:

> I'd like a little bit more explanation of what the final text size
 > is in the n.00-n.99 cases, with and without this modification.

Here are my additional findings:
While displaying text in the dialer name field, the code calculates maximum text size that can be fit in the field. Once it sends request to set that text size, the native code takes a floor of that number and set that size. Therefore, setting the text size between n.00 and n.99 for a given n does not impact the final width of the dialer name.

When this issue happens, the final width of the dailer name comes out to be bigger than the available text width. So it truncates. When LCD density is changes from 320dp to 360dp on this device, the issues goes away.

Patch-set: 4